# src/main/resources/api-definition.yml
swagger: '2.0'
info:
  description: Server to manage events for Beevy
  version: 1.0.0
  title: Beevy App
...
parameters:
...
  userID:
    name: userID
    in: path
    description: ID of a User
    required: true
    type: string
...
  tempAccessToken:
    name: tempAccessToken
    in: path
    description: tempAccessToken
    required: true
    type: string
paths:
...
  /event:
    get:
      tags:
        - event
      summary: Get all events that are available
      description: Get Events
      operationId: getEvents
      produces:
        - application/json
      responses:
        '200':
          description: successful operation
          schema:
            type: array
            items:
              $ref: '#/definitions/Event'
        '400':
          description: Failed to get events
  /event/create:
    post:
      tags:
        - event
      summary: Create an event
      description: Create an event for Beevy
      operationId: createEvent
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          description: Event Object
          schema:
            $ref: '#/definitions/Event'
      responses:
        '201':
          description: Created event
        '401':
          description: Unauthorized
        '405':
          description: Failed to create event
  /event/delete:
    post:
      tags:
        - event
      summary: Delete an event
      description: Delete an event from Beevy
      operationId: deleteEvent
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          description: Delete Event Object
          schema:
            $ref: '#/definitions/DeleteEventDTO'
      responses:
        '201':
          description: Deleted event
        '401':
          description: Unauthorized
        '405':
          description: Failed to delete event
  /event/join:
    post:
      tags:
        - event
      summary: Join an event
      description: Join an existing event
      operationId: joinEvent
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          description: Join Event Data
          schema:
            $ref: '#/definitions/JoinEventData'
      responses:
        '201':
          description: Joined event
        '405':
          description: Failed to join event
  /event/report:
    post:
      tags:
        - event
      summary: Report an event
      description: Report an event and send mail to admin team
      operationId: reportEvent
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          description: Report Event Data
          schema:
            $ref: '#/definitions/ReportDTO'
      responses:
        '201':
          description: Joined event
        '405':
          description: Failed to join event
  /event/{userID}/{tempAccessToken}:
      parameters:
        - $ref: '#/parameters/userID'
        - $ref: '#/parameters/tempAccessToken'
      get:
        tags:
          - event
        summary: Return a users events
        description: Get events of a user
        operationId: getUserEvents
        produces:
          - application/json
        responses:
          '200':
            description: successful operation
            schema:
              $ref: '#/definitions/UserEvents'
          '405':
            description: Failed to get events
...
definitions:
  Event:
    type: object
    properties:
      eventID:
        type: string
      admin:
        $ref: '#/definitions/MinimalUser'
      title:
        type: string
      summary:
        type: string
      description:
        type: string
      type:
        type: string
        enum:
          - project
          - activity
          - hangout
      date:
        type: string
      endDate:
        type: string
      address:
        $ref: '#/definitions/Address'
      tags:
        type: array
        items:
          type: string
      registeredMembers:
        type: array
        items:
          type: string
      possibleMemberCount:
        type: integer
      currentMemberCount:
        type: integer
  Address:
    type: object
    properties:
      street:
        type: string
      zip:
        type: integer
      city:
        type: string
      joinedEvents:
        type: array
        items:
          type: string
      createdEvents:
        type: array
        items:
          type: string
  MinimalUser:
    type: object
    properties:
      username:
        type: string
      userID:
        type: string
      token:
        type: string
      avatar:
        type: string
...
  ReportDTO:
    type: object
    properties:
      userID:
        type: string
      token:
        type: string
      eventID:
        type: string
      reason:
        type: string
  DeleteEventDTO:
    type: object
    properties:
      userID:
        type: string
      token:
        type: string
      eventID:
        type: string
...
  JoinEventData:
    type: object
    properties:
      userID:
        type: string
      token:
        type: string
      eventID:
        type: string
  UserEvents:
    type: object
    properties:
      joinedEvents:
        type: array
        items:
          $ref: '#/definitions/Event'
      createdEvents:
        type: array
        items:
          $ref: '#/definitions/Event'
...